Add a find_source hook for extension languages to locate missing source files.
This is implemented in Python as `gdb.find_source_hook`. If that property exists
and is callable it will be called with the path to the missing source file,
and if the callable returns a string that will be used as the new path
to the source file.

This can be used to fetch missing source files from a VCS or over HTTP,
for example.